+++ src/seamlesserverpkg/renderer/state.nim
@@ -1,0 +1,7 @@
+type
+  State* = ref object
+    ## State of frontend page
+
+func newState*: State =
+  ## Creates new state
+  new result